Create README.md
In this chess game project, I’m implementing various features such as the movement of chess pieces, detection of check and checkmate, and a simple graphical user interface for the game.
One of the key concepts I am learning while developing this project is object-oriented programming. Creating classes for each chess piece, such as the Pawn, Rook, Knight, Bishop, Queen, and King, and implementing their respective movement rules.
Another important concept I learnt is event-driven programming. I used the MouseListener interface to detect when a chess piece is clicked on and the MouseEvent class to get the coordinates of the click.
In addition, I also learnt how to use image icons in Java and how to display them in a JFrame using the JLabel class. I’ve also learnt about the GridLayout class and how to use it to create a chess board.
Overall, this project allowed me to gain a deeper understanding of Java programming and develop my skills in object-oriented programming, event-driven programming, and GUI development.